feat(parser): add grammar rules for function definitions and calls
This branch will introduce grammar support in the parser one rule per commit.
At each commit, we will implement one of the following grammar rules and remind what remains to be done in this list:

  function_definition: DEFINE function
  function: list_of_parameters block
  list_of_parameters: LPAREN parameters RPAREN
  parameters: | parameters atom_symbol
  block: LBRACE block_items RBRACE
  block_items: | block_items statement
  function_call: CALL atom_symbol list_of_numbers
  list_of_numbers: LPAREN atom_numbers RPAREN
  atom_numbers: | atom_numbers atom_number

Final goal: full grammar support for function definitions (with parameters and blocks) and function calls (with lists of numbers).